fix: consertando problema na visualização da imagem dos produtos

+++ src/main/resources/db/migration/V6__Create_user_product_table.sql
@@ -2,6 +2,6 @@ CREATE TABLE tb_products (
     id BIGSERIAL PRIMARY KEY,
     name VARCHAR(30) NOT NULL,
     link VARCHAR(255) NOT NULL,
-    price DECIMAL NOT NULL,
+    price FLOAT NOT NULL,
     image VARCHAR(255) NOT NULL
 );
